requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
ultralytics==8.3.139
pydantic==2.5.3
python-multipart==0.0.6
httpx==0.26.0
Pillow==10.1.0
numpy==1.26.3
transformers==4.52.4
opencv-python==4.8.1.78
aiofiles==23.2.1
python-jose==3.3.0
passlib==1.7.4
pydantic-settings==2.1.0
pydantic-core==2.14.6
anyio==4.7.0
websockets==12.0
python-dotenv==1.0.1
requests==2.31.0
typing-extensions==4.12.2
pyyaml==6.0.1
PytorchWildlife==1.2.3
zennit==0.5.1
# git+https://github.com/WildMeOrg/pairx.git#egg=pairx
# git+https://github.com/WildMeOrg/wbia-plugin-miew-id.git#egg=wbia-miew-id
